Update for variable Server path
Prior the Serverpath was hardcoded so the files had to be placed in the
root of the Webserver. Now it doesnt matter where they are placed. Added
a German translation. Added a fixed Murmur.php for ICE 3.5.1 +
Additional Notes in the README.md so you can create your own. Added some
stuff which wasnt translated yet.